LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-markers-focus-on-scrolling-expected.txt


FAIL active ::scroll-marker doesn't have focus on scroll if previous ::scroll-marker didn't have it assert_equals: ::scroll-marker opacity should be 1 but was  expected "1" but got ""
FAIL active ::scroll-marker saves focus on scroll if previous ::scroll-marker had it assert_equals: ::scroll-marker opacity should be 0.5 but was  expected "0.5" but got ""
